implement lazy loading for x509 certificates
The measurement table now has a `Kind` column which is equal to the
`Kind` of monitor used to generate that measurement. This will be useful
when deciding what type of information we might be able to lazy load for
a measurement.

The `Property` component will now display a `Chain` component for HTTP
measurements, which will fetch certificates when clicked.