use <div> instead of <dialog>
The <dialog> element will still be used for the modal dialog, but not the non-modal dialog. This is because the <dialog> element will enforce that the first button is focused upon opening, and since the buttons in this project use focus-visible, the user seems to "teleport" through the first button.

The dialog will render below the anchor in the dom, so using a div will avoid this issue and still work for keyboards. Popover API does the important work anyway.