implement `Repository` for `SQLiteRepository`
this includes a new `CommonRepository` which is not intended to fully implement the `Repository` interface, but just to contain functions that are compatible with all repository implementations.

If something is easier to write in a repository-specific way, it will just go in the actual implementation file. As pkg/sql evolves over time, more and more duplicate code will migrate to common.