apply dark defaults for dark themes
Theme names are appended to a new data-theme attribute and a default base color theme is chosen based on this value. When the theme name contains "dark", the dark defaults will be used. This makes it possible for themes to control which default theme they are layered over in the event that they want to override only a few color values.