final push toward theme system support across user interface
This removes generic CSS variables like --palette* because they are hard to define.

A default color palette is now included, and the new CSS variable system makes uses of these colors to build the default light and dark themes.

Most parts of the interface are easy to theme with the new variables, missing pieces can be implemented later or just targeted with standard CSS selectors.

Also includes a redesign of several interface elements including the monitor component, notably making it more compact for smaller screens and displaying last poll time, as well as a measurement visual aid for users that may not be able to identify the differences based on color.

Finally, updates README.md to include list of supported variables.