rewrite top level components, trim account context, add new hooks
Moved all of the effects out of the router and put them in Initialize or Dashboard depending on what they need to do. I think they were originally there because an early version had another route that wanted to share a lot of things with the dashboard.

Moved the websocket connection logic and "use*" functions into new hooks.